third_party/WebKit/LayoutTests/fast/forms/select-popup/popup-menu-open-partially-visible-expected.txt

PASS window.internals.pagePopupWindow.isWindowHidden() is false
PASS successfullyParsed is true

TEST COMPLETE
